\name{dosearch-package}
\alias{dosearch-package}
\alias{dosearch}
\docType{package}
\title{Causal Effect Identification from Multiple Incomplete Data Sources}

\description{
Solves causal effect identifiability problems from arbitrary observational and experimental data using a heuristic search. Allows for the presence of advanced data-generating mechanims.
}

\section{test}{text}

\author{Santtu Tikka, Antti Hyttinen, Juha Karvanen}

\references{
Bareinboim E., Pearl J. 2012 Causal Inference by Surrogate Experiments: z-identifiability. In \emph{Proceedings of the 28th Conference on Uncertainty in Artificial Intelligence}, 113--120.

Bareinboim E., Pearl J. 2013a A General Algorithm for Deciding Transportability of Experimental Results. \emph{Journal of Causal Inference}, \bold{1}, 107--134.

Bareinboim E., Pearl J. 2013b Meta-Transportability of Causal Effects: A Formal Approach. In \emph{Proceedings of the 16th International Conference on Artificial Intelligence and Statistics}, 135--143.

Bareinboim E., Pearl J. 2013c Causal Transportability with Limited Experiments. In \emph{Proceedings of the 27th AAAI Conference on Artificial Intelligence}, 95--101. 

Bareinboim E., Pearl J. 2014 Transportability from Multiple Environments with Limited Experiments: Completeness Results. \emph{Proceedings of the 27th Annual Conference on Neural Information Processing Systems}, 280--288.

Bareinboim E., Tian J. 2015 Recovering Causal Effects From Selection Bias. In \emph{Proceedings of the 29th AAAI Conference on Artificial Intelligence}, 3475--3481.

Mohan K., Pearl J., Tian, J. 2013 Graphical Models for Inference with Missing Data. In \emph{Advances in Neural Information Systems}, \bold{26}, 1277--1285.

Pearl J. 2009 \emph{Causality: Models, Reasoning and Inference}, New York: Cambridge University Press. 

Shpitser I., Pearl J. 2006a Identification of Joint Interventional Distributions in Recursive semi-Markovian Causal Models. In \emph{Proceedings of the 21st National Conference on Artificial Intelligence}, \bold{2}, 1219--1226. 

Shpitser I., Pearl J. 2006b Identification of Conditional Interventional Distributions. In \emph{Proceedings of the 22nd Conference on Uncertainty in Artificial Intelligence}, 427--444.

Tian J. 2002 Studies in Causal Reasoning and Learning. PhD thesis, Department of Computer Science, University of California, Los Angeles.

Tian, J., Pearl J. 2002 On Testable Implications of Causal Models with Hidden variables. In \emph{Proceedings of the Eighteenth Conference on Uncertainty in Artificial Intelligence}, 519--527.

Tikka, S., Karvanen J. 2017 Identifying Causal Effects with the R Package causaleffect. Journal of Statistical Software, \bold{76(12)}, 1--30.
} % END REFERENCES