% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dovalidate-package.R
\docType{package}
\name{dovalidate-package}
\alias{dovalidate-package}
\title{Causal Effect Identification from Multiple Incomplete Data Sources}
\description{
Solves causal effect identifiability problems from arbitrary
observational and experimental distributions using a heuristic search.
Allows for the presence of advanced data-generating mechanisms.
See Tikka et al. (2021) \url{doi:10.18637/jss.v099.i05} for further details.
}
\section{See also}{
\itemize{
\item The package vignette.
\item \code{\link[=dovalidate]{dovalidate()}} for instructions and various examples.
\item \url{https://github.com/santikka/dovalidate/issues/} to submit a bug report.
}
}

\references{
S. Tikka, A. Hyttinen and J. Karvanen.
"Causal effect identification from multiple incomplete data sources:
a general search-based approach." \emph{Journal of Statistical Software},
99(5):1--40, 2021.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikjuhva/dovalidate}
  \item Report bugs at \url{https://github.com/mikjuhva/dovalidate}
}

}
\author{
Santtu Tikka, Antti Hyttinen, Juha Karvanen
}
