% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{print.dosearch}
\alias{print.dosearch}
\title{Print a \code{dosearch} Object}
\usage{
\method{print}{dosearch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dosearch}.}

\item{...}{Additional arguments passed to \code{\link[base:format]{base::format()}}.}
}
\description{
Print a \code{dosearch} Object
}
\examples{
data <- "p(x,y,z)"
query <- "p(y|do(x))"
graph <- "
  x -> z
  Z -> y
  x <-> y
"
x <- dosearch(data, query, graph)
print(x)
}
