% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{print.summary.dosearch}
\alias{print.summary.dosearch}
\title{Print the Summary of a \code{dosearch} Object}
\usage{
\method{print}{summary.dosearch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.dosearch}.}

\item{...}{Not used.}
}
\description{
Print the Summary of a \code{dosearch} Object
}
