% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dovalidate.R
\name{validate_formula}
\alias{validate_formula}
\title{Validates does the given formula identify causal effect}
\usage{
validate_formula(cand_formula, query, graph)
}
\description{
Validates does the given formula identify causal effect
}
