% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{get_derivation}
\alias{get_derivation}
\title{Retrieve the Derivation of a Causal Query}
\usage{
get_derivation(x, run_again = FALSE, draw_all = FALSE)
}
\arguments{
\item{x}{An object of class \code{dosearch}.}

\item{run_again}{If \code{TRUE}, runs the search again in an attempt to obtain
the formula, for example if \code{control$formula} was \code{FALSE} in the call to
\code{\link[=dosearch]{dosearch()}}, but the query itself is identifiable.}

\item{draw_all}{A logical value. If \code{TRUE}, the derivation will contain
every step taken by the search. If \code{FALSE}, only steps that resulted in
identification are returned.}
}
\description{
Returns the derivation of a causal query of an object of class \code{dosearch}
returned by \code{\link[=dosearch]{dosearch()}}. If no derivation is available, returns
\code{NULL}.
}
\examples{
data <- "P(x,y,z)"
query <- "P(y|do(x))"
graph <- "
  x -> y
  z -> x
  z -> y
"
x <- dosearch(data, query, graph, control = list(draw_derivation = FALSE))
get_derivation(x, run_again = TRUE)
}
