% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{get_formula}
\alias{get_formula}
\title{Retrieve the Identifying Formula of a Causal Query}
\usage{
get_formula(x, run_again = FALSE)
}
\arguments{
\item{x}{An object of class \code{dosearch}.}

\item{run_again}{If \code{TRUE}, runs the search again in an attempt to obtain
the formula, for example if \code{control$formula} was \code{FALSE} in the call to
\code{\link[=dosearch]{dosearch()}}, but the query itself is identifiable.}
}
\value{
A \code{character} string representing the query in terms of the input
data or \code{NULL}.
}
\description{
Returns the identifying formula describing a causal query of an object of
class \code{dosearch} returned by \code{\link[=dosearch]{dosearch()}}. If no formula is
available, returns \code{NULL}.
}
\examples{
data <- "P(x,y,z)"
query <- "P(y|do(x))"
graph <- "
  x -> y
  z -> x
  z -> y
"
x <- dosearch(data, query, graph, control = list(formula = FALSE))
get_formula(x, run_again = TRUE)
}
