% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{is_identifiable}
\alias{is_identifiable}
\title{Was the Target Distribution Identifiable?}
\usage{
is_identifiable(x)
}
\arguments{
\item{x}{An object of class \code{dosearch}.}
}
\value{
A logical value. If \code{TRUE}, the target distribution was identifiable
from the available inputs.
}
\description{
Returns the a logical value describing the identifiability of a causal query
of an object of class \code{dosearch} returned by \code{\link[=dosearch]{dosearch()}}.
}
\examples{
data <- "P(x,y,z)"
query <- "P(y|do(x))"
graph <- "
  x -> y
  z -> x
  z -> y
"
x <- dosearch(data, query, graph)
is_identifiable(x)
# TRUE
}
