% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosearch.R
\name{summary.dosearch}
\alias{summary.dosearch}
\title{Summary of a \code{dosearch} Object}
\usage{
\method{summary}{dosearch}(object, ...)
}
\arguments{
\item{object}{An object of class \code{dosearch}.}

\item{...}{Not used.}
}
\description{
Summary of a \code{dosearch} Object
}
